+++ frontend/src/app/components/common/tooltip-content.tsx
